Added a second Activity and moved the Settings there.
