+++ app/src/main/res/values-en/strings.xml
@@ -40,4 +40,5 @@
     <string name="app_name">JulyTimer</string>
     <string name="missed_percent_2_singular">Percent jump. Very Sad.</string>
     <string name="missed_percent_2_plural">Percent jumps. Very Sad.</string>
+    <string name="title_activity_settings">Settings</string>
 </resources>
