+++ app/src/main/res/values/strings.xml
@@ -47,5 +47,7 @@
     <string name="milestone_almost_done_toast">Fast alles geschafft! Bis bald :)</string>
     <string name="milestone_month_toast">Noch einen Monat geschafft! Ich bin stolz auf dich!</string>
     <string name="milestone_done_toast">Jaaaaaaa!</string>
+    <string name="title_activity_settings">Einstellungen</string>
+    <!-- Strings used for fragments for navigation -->
 
 </resources>
